rework started. dump project restoration

+++ classes/App/Controller/Ajax.php
@@ -53,22 +53,89 @@ class Ajax extends \PHPixie\Controller {
         if(!$this->logged_in('admin'))
             return;
 
+//        Session::set('inActual',true);
+
         $id = $this->request->param("id");
         if (!$id) {
             return;
         }
         $a = $this->pixie->orm->get('award')->where('id',$id)->find();
 
-        $a->delete();
+        if ($a->loaded()) {
+            $d= $a->year;
+            $ds= $a->stage_id;
+            $a->delete();
+            $_SESSION['dirty_year'] = $d;
+            $_SESSION['dirty_stage'] = $ds;
+        }
     }
 
+    /**
+     * удаляет указанный рассчет для пользователя
+     */
+    public function action_delete_awarduser() {
+        if(!$this->logged_in('admin'))
+            return;
+
+        $id = $this->request->param("id");
+        if (!$id) {
+            return;
+        }
+        $a = $this->pixie->orm->get('awarduser')->where('id',$id)->find();
+
+        if ($a->loaded()) {
+            $d= $a->year;
+            $ds= $a->stage_id;
+            $a->delete();
+            $_SESSION['dirty_year'] = $d;
+            $_SESSION['dirty_stage'] = $ds;
+        }
+
+    }
+
+    /**
+     * проверяет коичество расчетов на выбранный факультет, Если их нет то печалька
+     */
+    public function action_check_funduser_calc_count() {
+        if(!$this->logged_in('admin'))
+            return;
+        if ($this->request->method == 'GET'){
+            $fac = $this->request->get('id');
+            $year = $this->request->get('year');
+            $stage_id = $this->request->get('stage');
+            $qwe = $this->pixie->orm->get('awarduser')->with('user.faculty')->where('year',$year)->where('stage_id',$stage_id)->where('a1.id',$fac)->find_all()->as_array();
+
+            echo count($qwe);
+        }
+    }
+
+
+    public function action_get_fac_info() {
+        if(!$this->logged_in('admin'))
+            return;
+
+        $id = $this->request->get("id");
+        $nf = 0;
+        $nprf = 0;
+        if ($id != null) {
+            $fac = $this->pixie->orm->get('faculty')->where('id',$id)->find();
+            $nf = $fac->nf;
+            $nprf = $fac->nprf;
+        }
+        echo json_encode(array(
+            'nprf'=>$nprf,
+            'nf'=>$nf
+        ));
+
+
+    }
 
     protected function logged_in($role = null){
         if($this->pixie->auth->user() == null){
             return false;
         }
 
-        if($role && !$this->pixie->auth->has_role($role)){
+        if(!$this->pixie->auth->has_role('super') && $role && !$this->pixie->auth->has_role($role)){
             return false;
         }
